python_assignment/requirement.txt
kafka-python==2.0.1
numpy==1.19.0
pandas==1.0.5
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
six==1.15.0
websockets==8.1
yapf==0.30.0
